+++ agate/csv_py2.py
@@ -250,7 +250,8 @@ class Sniffer(object):
         """
         try:
             dialect = csv.Sniffer().sniff(sample, POSSIBLE_DELIMITERS)
-        except csv.Error:
+        except csv.Error as e:
+            warnings.warn('Error sniffing CSV dialect: %s' % e, RuntimeWarning, stacklevel=2)
             dialect = None
 
         return dialect